sdks/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "slop-security"
version = "1.0.0"
description = "One-line security for AI-generated code. OWASP Top 10 protection out of the box."
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "Slop Security Team", email = "team@slopsecurity.io" }
]
keywords = [
    "security",
    "owasp",
    "xss",
    "sql-injection",
    "csrf",
    "middleware",
    "flask",
    "fastapi",
    "django",
    "slop",
    "vibecoding",
    "ai-generated"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Middleware",
    "Framework :: Flask",
    "Framework :: FastAPI",
    "Framework :: Django",
]
dependencies = [
    "argon2-cffi>=21.0.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
flask = ["flask>=2.0.0"]
fastapi = ["fastapi>=0.100.0", "starlette>=0.27.0"]
django = ["django>=4.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://slopsecurity.io"
Documentation = "https://slopsecurity.io/docs/python"
Repository = "https://github.com/slop-security/slop-security"
Issues = "https://github.com/slop-security/slop-security/issues"

[tool.hatch.build.targets.wheel]
packages = ["slop"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "S", "B"]

[tool.mypy]
python_version = "3.8"
strict = true
